refactor: remove 'refactored' suffix from service and test files
## Summary
Renamed all service files with 'refactored' suffix to use standard naming conventions.
This improves code clarity by removing historical context from filenames.

## Changes
**Service Files:**
- feedback_service_refactored.py → feedback_service.py
- llm_service_refactored.py → llm_service.py
- session_service_refactored.py → session_service.py

**Test Files:**
- test_feedback_service_refactored.py → test_feedback_service.py
- test_llm_service_refactored.py → test_llm_service.py
- test_session_service_refactored.py → test_session_service.py

**Updated Imports:**
- app/roleplaying/services/dependencies/feedback.py (4 imports)
- app/roleplaying/services/dependencies/llm.py (7 imports)
- app/roleplaying/services/dependencies/services.py (1 import)
- Updated docstrings in llm_service.py
