+++ app/roleplaying/services/business/session_service.py
